+++ CHANGELOG.md
@@ -62,5 +62,5 @@
 
 # 0.1.7
 * Define `ifelse` function for `Boolean` values
-* Change `Stack` and `Nonempty` definitions, temporarily remove `filter` and `linearize`
+* Change `Stack` and `Nonempty` definitions, temporarily remove `filter`
 * Define `<**>`, `<***>`, `<****>` methods for `Applicative` to compact expressions